Runtime/r3/linux: Initial custom native init implementation which registers a signal handler for SIGSEGV/SIGABRT/SIGBUS trying to log information about a crash to the default log, bugref:11017 [fix]
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@111428 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/r3/linux/init-linux.cpp
@@ -268,7 +268,7 @@ static void rtR3LnxSigSegvBusHandler(int iSignum, siginfo_t *pSigInfo, void *pvC
                 aszCmdline[cbRead] = '\0'; /* Terminate */
                 RTLogLoggerWeak(pLogger, NULL, "\nCommandLine: ");
 
-                uint32_t off = 0;
+                ssize_t off = 0;
                 while (off < cbRead)
                 {
                     RTLogLoggerWeak(pLogger, NULL, "%s\n", &aszCmdline[off]);
